+++ src/Oro/Bundle/LayoutBundle/DependencyInjection/Configuration.php
@@ -58,7 +58,7 @@ class Configuration implements ConfigurationInterface
     /**
      * Appends config nodes for "templating"
      */
-    protected function appendTemplatingNodes(ArrayNodeDefinition $parentNode)
+    protected function appendTemplatingNodes(ArrayNodeDefinition $parentNode): void
     {
         $treeBuilder = new TreeBuilder('templating');
         $node        = $treeBuilder->getRootNode();
@@ -98,7 +98,7 @@ class Configuration implements ConfigurationInterface
     /**
      * Appends config nodes for "themes"
      */
-    protected function appendThemingNodes(ArrayNodeDefinition $parentNode)
+    protected function appendThemingNodes(ArrayNodeDefinition $parentNode): void
     {
         $parentNode
             ->children()
@@ -108,6 +108,12 @@ class Configuration implements ConfigurationInterface
                 ->end()
                 ->scalarNode('active_theme')
                     ->info('The identifier of the theme that should be used by default')
+                ->end()
+                ->arrayNode('inherited_theme_options')
+                    ->info('List of inherited theme options or theme config options.')
+                    ->example(['svg_icons_support', 'config.icons'])
+                    ->prototype('scalar')->end()
+                    ->defaultValue([])
                 ->end()
             ->end();
     }