+++ src/Oro/Bundle/LayoutBundle/Layout/Block/Type/FormType.php
@@ -93,7 +93,7 @@ class FormType extends AbstractType
 
         $options = $options->toArray();
         foreach ($options['additional_block_prefixes'] as &$blockPrefix) {
-            $blockPrefix .=  $suffix;
+            $blockPrefix .= $suffix;
         }
         unset($blockPrefix);