+++ src/Oro/Component/Layout/Extension/Theme/Model/ThemeFactory.php
@@ -52,6 +52,7 @@ class ThemeFactory implements ThemeFactoryInterface
             'directory',
             'groups',
             'description',
+            'fonts'
         ];
 
         foreach ($properties as $property) {